Bug fixed broken video when switch to non-iframe allowed site.
This bug is discovered during the iframe is played within one place and when switch tabs to pages/ domain such as facebook or github. That it doesnt allow injections of iframe, while the youtube Iframe API still send messages to the html. But the message is trying to init the video, yet provide videoID as null or undefined. This bug is when videoID us === null, which is patched by conditional constrainting the if statment.