Bug fixed for display during refeash of the search results is squashed to the left.
This bug was discovered by refeash tabs between the window, which the search results will be collpased to the right due to the lack of repaint. This is cause by the display is not shown intitially for the browser to repaint, so mouse on and leave listeners are attached and deattached dynamically and display is set to show after the visionbility is hidden (user leaving the page) for future repainting.